+++ src/test/java/com/gft/productprices/prices/infrastructure/persistence/mapper/PriceEntityMapperTest.java
@@ -1,8 +1,5 @@
 package com.gft.productprices.prices.infrastructure.persistence.mapper;
-import com.gft.productprices.prices.domain.model.Brand;
-import com.gft.productprices.prices.domain.model.DateRange;
 import com.gft.productprices.prices.domain.model.Price;
-import com.gft.productprices.prices.domain.model.ProductPrice;
 import com.gft.productprices.prices.infrastructure.persistence.entities.PriceEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -30,9 +27,8 @@ class PriceEntityMapperTest {
 
     @Test
     void testEntityToDomainMapping() {
-        // Mock data
         int priceList = 1;
-        long productId = 100L;
+        int productId = 100;
         int priority = 2;
         double price = 50.0;
         String currency = "USD";
@@ -40,7 +36,6 @@ class PriceEntityMapperTest {
         Date startDate = new Date();
         Date endDate = new Date();
 
-        // Set up the mock PriceEntity
         when(priceEntity.getPriceList()).thenReturn(priceList);
         when(priceEntity.getProductId()).thenReturn(productId);
         when(priceEntity.getPriority()).thenReturn(priority);
@@ -50,10 +45,8 @@ class PriceEntityMapperTest {
         when(priceEntity.getStartDate()).thenReturn(startDate);
         when(priceEntity.getEndDate()).thenReturn(endDate);
 
-        // Map PriceEntity to Price using the mapper
         Price mappedPrice = priceEntityMapper.entityToDomain(priceEntity);
 
-        // Verify mapping
         assertEquals(priceList, mappedPrice.getPriceListId());
         assertEquals(productId, mappedPrice.getProductId());
         assertEquals(priority, mappedPrice.getPriority());
